Refactor http2 client handlers
