+++ components/org.wso2.transport.http.netty/src/main/java/org/wso2/transport/http/netty/message/Http2Frame.java
@@ -1,0 +1,44 @@
+/*
+ * Copyright (c) 2018, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ * WSO2 Inc. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.wso2.transport.http.netty.message;
+
+/**
+ * {@code Http2Frame} is the base for all HTTP/2 frames.
+ */
+public abstract class Http2Frame {
+
+    private int streamId;
+    private boolean endOfStream;
+
+    public int getStreamId() {
+        return streamId;
+    }
+
+    public void setStreamId(int streamId) {
+        this.streamId = streamId;
+    }
+
+    public boolean isEndOfStream() {
+        return endOfStream;
+    }
+
+    public void setEndOfStream(boolean endOfStream) {
+        this.endOfStream = endOfStream;
+    }
+}
