+++ components/org.wso2.transport.http.netty/src/main/java/org/wso2/transport/http/netty/sender/http2/Http2TargetHandler.java
@@ -1,32 +1,38 @@
 /*
- *   Copyright (c) 2018, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ * Copyright (c) 2018, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
  *
- *   WSO2 Inc. licenses this file to you under the Apache License,
- *   Version 2.0 (the "License"); you may not use this file except
- *   in compliance with the License.
- *   You may obtain a copy of the License at
+ * WSO2 Inc. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ *    http://www.apache.org/licenses/LICENSE-2.0
  *
- *  Unless required by applicable law or agreed to in writing,
- *  software distributed under the License is distributed on an
- *  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- *  KIND, either express or implied.  See the License for the
- *  specific language governing permissions and limitations
- *  under the License.
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
  */
 
 package org.wso2.transport.http.netty.sender.http2;
 
 import io.netty.buffer.ByteBuf;
+import io.netty.channel.ChannelDuplexHandler;
 import io.netty.channel.ChannelHandlerContext;
-import io.netty.channel.ChannelOutboundHandlerAdapter;
 import io.netty.channel.ChannelPromise;
+import io.netty.handler.codec.http.DefaultHttpContent;
+import io.netty.handler.codec.http.DefaultHttpResponse;
+import io.netty.handler.codec.http.DefaultLastHttpContent;
 import io.netty.handler.codec.http.EmptyHttpHeaders;
 import io.netty.handler.codec.http.HttpContent;
 import io.netty.handler.codec.http.HttpHeaders;
 import io.netty.handler.codec.http.HttpMessage;
 import io.netty.handler.codec.http.HttpRequest;
+import io.netty.handler.codec.http.HttpResponse;
+import io.netty.handler.codec.http.HttpResponseStatus;
+import io.netty.handler.codec.http.HttpVersion;
 import io.netty.handler.codec.http.LastHttpContent;
 import io.netty.handler.codec.http2.EmptyHttp2Headers;
 import io.netty.handler.codec.http2.Http2CodecUtil;
@@ -41,23 +47,29 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.wso2.transport.http.netty.common.Constants;
 import org.wso2.transport.http.netty.common.Util;
+import org.wso2.transport.http.netty.message.DefaultListener;
 import org.wso2.transport.http.netty.message.HTTPCarbonMessage;
+import org.wso2.transport.http.netty.message.Http2DataFrame;
+import org.wso2.transport.http.netty.message.Http2HeadersFrame;
+import org.wso2.transport.http.netty.message.Http2PushPromise;
 import org.wso2.transport.http.netty.message.Http2Reset;
+import org.wso2.transport.http.netty.message.HttpCarbonResponse;
+import org.wso2.transport.http.netty.message.PooledDataStreamerFactory;
 
 import java.util.Locale;
 
 /**
- * {@code ClientOutboundHandler} is responsible for writing HTTP/2 Frames to the backend service.
+ * {@code Http2TargetHandler} is responsible for sending and receiving HTTP/2 frames over a connection.
  */
-public class ClientOutboundHandler extends ChannelOutboundHandlerAdapter {
+public class Http2TargetHandler extends ChannelDuplexHandler {
 
-    private static final Logger log = LoggerFactory.getLogger(ClientOutboundHandler.class);
+    private static final Logger log = LoggerFactory.getLogger(Http2TargetHandler.class);
     private Http2Connection connection;
     // Encoder associated with the HTTP2ConnectionHandler
     private Http2ConnectionEncoder encoder;
     private Http2ClientChannel http2ClientChannel;
 
-    public ClientOutboundHandler(Http2Connection connection, Http2ConnectionEncoder encoder) {
+    public Http2TargetHandler(Http2Connection connection, Http2ConnectionEncoder encoder) {
         this.connection = connection;
         this.encoder = encoder;
     }
@@ -66,7 +78,7 @@ public class ClientOutboundHandler extends ChannelOutboundHandlerAdapter {
     public void write(ChannelHandlerContext ctx, Object msg, ChannelPromise promise) throws Exception {
         if (msg instanceof OutboundMsgHolder) {
             OutboundMsgHolder outboundMsgHolder = (OutboundMsgHolder) msg;
-            new Http2RequestWriter(outboundMsgHolder).writeContent(ctx);
+            new Http2TargetHandler.Http2RequestWriter(outboundMsgHolder).writeContent(ctx);
         } else if (msg instanceof Http2Reset) {
             Http2Reset resetMsg = (Http2Reset) msg;
             resetStream(ctx, resetMsg.getStreamId(), resetMsg.getError());
@@ -268,5 +280,145 @@ public class ClientOutboundHandler extends ChannelOutboundHandlerAdapter {
         http2ClientChannel.getDataEventListeners().
                 forEach(dataEventListener -> dataEventListener.onStreamReset(streamId));
         ctx.flush();
+    }
+
+    public void channelRead(ChannelHandlerContext ctx, Object msg) throws Exception {
+        if (msg instanceof Http2HeadersFrame) {
+            onHeadersRead(ctx, (Http2HeadersFrame) msg);
+        } else if (msg instanceof Http2DataFrame) {
+            onDataRead((Http2DataFrame) msg);
+        } else if (msg instanceof Http2PushPromise) {
+            onPushPromiseRead((Http2PushPromise) msg);
+        }
+    }
+
+    public void onHeadersRead(ChannelHandlerContext ctx, Http2HeadersFrame http2HeadersFrame) throws Http2Exception {
+
+        int streamId = http2HeadersFrame.getStreamId();
+        boolean endStream = http2HeadersFrame.isEndOfStream();
+        Http2Headers headers = http2HeadersFrame.getHeaders();
+
+        OutboundMsgHolder outboundMsgHolder = http2ClientChannel.getInFlightMessage(streamId);
+        boolean isServerPush = false;
+        if (outboundMsgHolder == null) {
+            outboundMsgHolder = http2ClientChannel.getPromisedMessage(streamId);
+            if (outboundMsgHolder != null) {
+                isServerPush = true;
+            } else {
+                log.warn("Header Frame received on channel: {} with invalid stream id: {} ",
+                         http2ClientChannel.toString(), streamId);
+                return;
+            }
+        }
+        // Create response carbon message
+        HttpCarbonResponse responseMessage = setupResponseCarbonMessage(ctx, streamId, headers, outboundMsgHolder);
+        if (isServerPush) {
+            outboundMsgHolder.addPushResponse(streamId, responseMessage);
+            if (endStream) {
+                responseMessage.addHttpContent(new DefaultLastHttpContent());
+                http2ClientChannel.removePromisedMessage(streamId);
+            }
+        } else {
+            if (endStream) {
+                responseMessage.addHttpContent(new DefaultLastHttpContent());
+                http2ClientChannel.removeInFlightMessage(streamId);
+            }
+            outboundMsgHolder.setResponse(responseMessage);
+        }
+    }
+
+    public void onDataRead(Http2DataFrame dataFrame) throws Http2Exception {
+        int streamId = dataFrame.getStreamId();
+        ByteBuf data = dataFrame.getData();
+        boolean endOfStream = dataFrame.isEndOfStream();
+
+        OutboundMsgHolder outboundMsgHolder = http2ClientChannel.getInFlightMessage(streamId);
+        boolean isServerPush = false;
+        if (outboundMsgHolder == null) {
+            outboundMsgHolder = http2ClientChannel.getPromisedMessage(streamId);
+            if (outboundMsgHolder != null) {
+                isServerPush = true;
+            } else {
+                log.warn("Data Frame received on channel: {} with invalid stream id: {}",
+                         http2ClientChannel.toString(), streamId);
+                return;
+            }
+        }
+        if (isServerPush) {
+            HTTPCarbonMessage responseMessage = outboundMsgHolder.getPushResponse(streamId);
+            if (endOfStream) {
+                responseMessage.addHttpContent(new DefaultLastHttpContent(data.retain()));
+                http2ClientChannel.removePromisedMessage(streamId);
+            } else {
+                responseMessage.addHttpContent(new DefaultHttpContent(data.retain()));
+            }
+        } else {
+            HTTPCarbonMessage responseMessage = outboundMsgHolder.getResponse();
+            if (endOfStream) {
+                responseMessage.addHttpContent(new DefaultLastHttpContent(data.retain()));
+                http2ClientChannel.removeInFlightMessage(streamId);
+            } else {
+                responseMessage.addHttpContent(new DefaultHttpContent(data.retain()));
+            }
+        }
+    }
+
+    public void onPushPromiseRead(Http2PushPromise pushPromise) throws Http2Exception {
+
+        int streamId = pushPromise.getStreamId();
+        int promisedStreamId = pushPromise.getPromisedStreamId();
+
+        if (log.isDebugEnabled()) {
+            log.debug("Received a push promise on channel: {} over stream id: {}, promisedStreamId: {}",
+                      http2ClientChannel.toString(), streamId, promisedStreamId);
+        }
+
+        OutboundMsgHolder outboundMsgHolder = http2ClientChannel.getInFlightMessage(streamId);
+        if (outboundMsgHolder == null) {
+            log.warn("Push promise received in channel: {} over invalid stream id : {}",
+                     http2ClientChannel.toString(), streamId);
+            return;
+        }
+        http2ClientChannel.putPromisedMessage(promisedStreamId, outboundMsgHolder);
+        pushPromise.setOutboundMsgHolder(outboundMsgHolder);
+        outboundMsgHolder.addPromise(pushPromise);
+    }
+
+    private HttpCarbonResponse setupResponseCarbonMessage(ChannelHandlerContext ctx, int streamId,
+                                                          Http2Headers http2Headers,
+                                                          OutboundMsgHolder outboundMsgHolder) {
+        // Create HTTP Response
+        CharSequence status = http2Headers.status();
+        HttpResponseStatus responseStatus;
+        try {
+            responseStatus = HttpConversionUtil.parseStatus(status);
+        } catch (Http2Exception e) {
+            responseStatus = HttpResponseStatus.BAD_GATEWAY;
+        }
+        HttpVersion version = new HttpVersion(Constants.HTTP_VERSION_2_0, true);
+        HttpResponse httpResponse = new DefaultHttpResponse(version, responseStatus);
+
+        // Set headers
+        try {
+            HttpConversionUtil.addHttp2ToHttpHeaders(
+                    streamId, http2Headers, httpResponse.headers(), version, false, false);
+        } catch (Http2Exception e) {
+            outboundMsgHolder.getResponseFuture().
+                    notifyHttpListener(new Exception("Error while setting http headers", e));
+        }
+        // Create HTTP Carbon Response
+        HttpCarbonResponse responseCarbonMsg = new HttpCarbonResponse(httpResponse, new DefaultListener(ctx));
+
+        // Setting properties of the HTTP Carbon Response
+        responseCarbonMsg.setProperty(Constants.POOLED_BYTE_BUFFER_FACTORY,
+                                      new PooledDataStreamerFactory(ctx.alloc()));
+        responseCarbonMsg.setProperty(Constants.DIRECTION, Constants.DIRECTION_RESPONSE);
+        responseCarbonMsg.setProperty(Constants.HTTP_STATUS_CODE, httpResponse.status().code());
+
+        /* copy required properties for service chaining from incoming carbon message to the response carbon message
+        copy shared worker pool */
+        responseCarbonMsg.setProperty(Constants.EXECUTOR_WORKER_POOL,
+                                      outboundMsgHolder.getRequest().getProperty(Constants.EXECUTOR_WORKER_POOL));
+        return responseCarbonMsg;
     }
 }
